refactor: added one implementation of fibonacciCahce
